Fix function not found error
