+++ helpers/formatData.js
@@ -1,6 +1,3 @@
-const config = require('../config');
-const axios = require('axios');
-
 function putVehiclesIntoRoutes(vehicles) {
     const routes = {};
     vehicles.forEach((vehicle) => {
@@ -14,53 +11,5 @@ function putVehiclesIntoRoutes(vehicles) {
     }));
 }
 
-function getRoutes() {
-    return axios.get('/agencies/sf-muni/routes', {
-      baseURL: config.restbusURL
-    })
-    .then((response) => {
-      const routes = response.data.id;
-      console.log(routes);
-      return routes;
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-}
 
-function getStopsFromRoute(routes) {
-    Promise.all(routes.map(route => {
-      return axios.get(`/agencies/sf-muni/routes/${route}`, {
-      baseURL: config.restbusURL
-      })
-    }))
-    .then((responses) => {
-      return responses.map(response => {  
-        const stops = response.data.stops;
-        return stops.map(addStopsToRoute).map(stopObj => {
-            stopObj.rid = response.data.id;
-            return stopObj;
-        });
-      })
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-}
-
-function addStopsToRoute(nextbusObject) {
-  const { id, name, lat, lon } = nextbusObject;
-  return {
-    sid: id,
-    title: name,
-    lat,
-    lon,
-  };
-}
-
-module.exports = {
-  putVehiclesIntoRoutes,
-  getRoutes,
-  getStopsFromRoute,
-  addStopsToRoute
-}
+module.exports = putVehiclesIntoRoutes;
